# 55% memchr optimization with SIMD 