--- a/imgui.cpp
+++ b/imgui.cpp
@@ -1943,6 +1943,107 @@ ImVec2 ImTriangleClosestPoint(const ImVec2& a, const ImVec2& b, const ImVec2& c,
 // [SECTION] MISC HELPERS/UTILITIES (String, Format, Hash functions)
 //-----------------------------------------------------------------------------
 
+#if defined IMGUI_ENABLE_AVX2_IMMEMCHR
+const void* ImMemchr(const void* buf, int val, size_t count)
+{
+    const size_t SIMD_LENGTH = 32;
+    const size_t SIMD_LENGTH_MASK = SIMD_LENGTH - 1;
+
+    const unsigned char* ptr = (const unsigned char*)buf;
+    const unsigned char* end = ptr + count;
+    const unsigned char* align_end = end - SIMD_LENGTH;
+    const unsigned char ch = (const unsigned char)val;
+
+    if (ptr <= align_end)
+    {
+        const __m256i target = _mm256_set1_epi8(ch);
+
+        if ((uintptr_t)ptr & SIMD_LENGTH_MASK)
+        {
+            __m256i chunk = _mm256_lddqu_si256((const __m256i*)ptr);
+            int mask = _mm256_movemask_epi8(_mm256_cmpeq_epi8(chunk, target));
+
+            if (mask)
+                return (const void*)(ptr + _tzcnt_u32(mask));
+
+            ptr = (const unsigned char*)_andn_u64(SIMD_LENGTH_MASK, (uintptr_t)ptr + SIMD_LENGTH_MASK);
+        }
+
+        for (; ptr <= align_end; ptr += SIMD_LENGTH)
+        {
+            __m256i chunk = _mm256_load_si256((const __m256i*)ptr);
+            int mask = _mm256_movemask_epi8(_mm256_cmpeq_epi8(chunk, target));
+
+            if (mask)
+                return (const void*)(ptr + _tzcnt_u32(mask));
+
+            if (ptr <= end - 1024)
+                _mm_prefetch((const char*)(ptr + 1024), _MM_HINT_T0);
+        }
+    }
+
+    for (; ptr < end; ptr++)
+    {
+        if (*ptr == ch)
+            return (const void*)(ptr);
+    }
+
+    return nullptr;
+}
+#elif defined IMGUI_ENABLE_SSE_IMMEMCHR
+const void* ImMemchr(const void* buf, int val, size_t count)
+{
+    const size_t SIMD_LENGTH = 16;
+    const size_t SIMD_LENGTH_MASK = SIMD_LENGTH - 1;
+
+    const unsigned char* ptr = (const unsigned char*)buf;
+    const unsigned char* end = ptr + count;
+    const unsigned char* align_end = end - SIMD_LENGTH;
+    const unsigned char ch = (const unsigned char)val;
+
+    if (ptr <= align_end)
+    {
+        const __m128i target = _mm_set1_epi8(ch);
+
+        if ((uintptr_t)ptr & SIMD_LENGTH_MASK)
+        {
+            __m128i chunk = _mm_lddqu_si128((const __m128i*)ptr);
+            int mask = _mm_movemask_epi8(_mm_cmpeq_epi8(chunk, target));
+
+            if (mask)
+                return (const void*)(ptr + _tzcnt_u32(mask));
+
+            ptr = (const unsigned char*)(((uintptr_t)ptr + SIMD_LENGTH_MASK) & ~SIMD_LENGTH_MASK);
+        }
+
+        for (; ptr <= align_end; ptr += SIMD_LENGTH)
+        {
+            __m128i chunk = _mm_load_si128((const __m128i*)ptr);
+            int mask = _mm_movemask_epi8(_mm_cmpeq_epi8(chunk, target));
+
+            if (mask)
+                return (const void*)(ptr + _tzcnt_u32(mask));
+
+            if (ptr <= end - 1024)
+                _mm_prefetch((const char*)(ptr + 1024), _MM_HINT_T0);
+        }
+    }
+
+    for (; ptr < end; ptr++)
+    {
+        if (*ptr == ch)
+            return (const void*)(ptr);
+    }
+
+    return nullptr;
+}
+#else
+const void* ImMemchr(const void* buf, int val, size_t count)
+{
+    return memchr(buf, val, count);
+}
+#endif
+
 // Consider using _stricmp/_strnicmp under Windows or strcasecmp/strncasecmp. We don't actually use either ImStricmp/ImStrnicmp in the codebase any more.
 int ImStricmp(const char* str1, const char* str2)
 {
@@ -1990,7 +2091,7 @@ char* ImStrdupcpy(char* dst, size_t* p_dst_size, const char* src)
 
 const char* ImStrchrRange(const char* str, const char* str_end, char c)
 {
-    const char* p = (const char*)memchr(str, (int)c, str_end - str);
+    const char* p = (const char*)ImMemchr(str, (int)c, str_end - str);
     return p;
 }
 
@@ -2005,7 +2106,7 @@ int ImStrlenW(const ImWchar* str)
 // Find end-of-line. Return pointer will point to either first \n, either str_end.
 const char* ImStreolRange(const char* str, const char* str_end)
 {
-    const char* p = (const char*)memchr(str, '\n', str_end - str);
+    const char* p = (const char*)ImMemchr(str, '\n', str_end - str);
     return p ? p : str_end;
 }
 
@@ -2554,7 +2655,7 @@ int ImTextCountLines(const char* in_text, const char* in_text_end)
     int count = 0;
     while (in_text < in_text_end)
     {
-        const char* line_end = (const char*)memchr(in_text, '\n', in_text_end - in_text);
+        const char* line_end = (const char*)ImMemchr(in_text, '\n', in_text_end - in_text);
         in_text = line_end ? line_end + 1 : in_text_end;
         count++;
     }
@@ -2962,7 +3063,7 @@ void ImGuiTextIndex::append(const char* base, int old_size, int new_size)
     if (EndOffset == 0 || base[EndOffset - 1] == '\n')
         LineOffsets.push_back(EndOffset);
     const char* base_end = base + new_size;
-    for (const char* p = base + old_size; (p = (const char*)memchr(p, '\n', base_end - p)) != 0; )
+    for (const char* p = base + old_size; (p = (const char*)ImMemchr(p, '\n', base_end - p)) != 0; )
         if (++p < base_end) // Don't push a trailing offset on last \n
             LineOffsets.push_back((int)(intptr_t)(p - base));
     EndOffset = ImMax(EndOffset, new_size);