@Kraionix 2025-02-21
Hello.
I am learning low-level optimizations using SIMD on x86-64. I wrote the code from scratch for the sake of practice.
I optimized the SSE and AVX2 implementations of ImMemchr to minimize clock cycles as much as possible.
I took into account loading unaligned data, preloading data into the cache, I also took security into account.
I optimized iteratively until I got to the current code, with maximum optimization.
This is the benchmark source code and binary [link](https://github.com/Kraionix/ImGuiMemchrBench).