@Kraionix 2025-02-21
I also tested `ImGui::TextUnformatted` with a 16 MB `std::string` buffer, in which every 131 characters is `\n`, the performance on AVX2 ImMemchr increased by about 60%, compared to regular memchr. `ImGui::TextUnformatted` call was 1.6 ms per frame, became 0.98 ms.