@Kraionix 2025-03-01
I added an optimized ImStrlen with implementations on SSE and AVX2.