@DominiquePelle-TomTom 2025-03-01
I'm very skeptical that we should replace the std functions such as `strlen()` or `strchr()`.

If they can be optimized, then they should be optimized in the C library, not in the ImGui library.